Create add new entry endpoint
API endpoint for new entry creation; it takes a request payload in JSON
format which should contain; title, description, and conclusion of entry
to be created.

Title and description of entry to be created are required and if empty,
 the endpoint responds to the request with a status code value, 400 and
a JSON response data of status and data, with status set to failed and
data set to an object containing error message.

With correct request payload, a new entry will be created and the
endpoint will respond with status code value, 201 and a JSON reponse
data of status and data, with status set to succeeded and data set to an
object containing success message.