Create fetch all entries endpoint
API endpoint for fetching all entries in the entry list. It takes no
request payload, and should return response data in JSON format
containing string and array identifiers; status and data respectively,
to the requesting client, with status set to succeeded and data set to
array of all entries. It should also return to the client, a HTTP success
status code, 200.