Create fetch single entry endpoint
API endpoint for fetching single entry from entry list using it's
position Id (index). The Id should be append to the end of request url
with a preceeding forward slash, (eg. /api/v1/entries/Id).

If Id is not a valid non-negative integer value, the endpoint returns a
JSON error object and a HTTP status code of 400, to the requesting client.

If Id is valid and an entry object found at that position in the entry
list, the endpoint returns a JSON object containing the entry and a HTTP
status code of 200, to the client. Else, the endpoint returns a "Not
found" message in JSON and a HTTP status code of 404, to the client.