Create modify entry endpoint
API endpoint for modifying saved entry using it's position Id (index) in
the entry list. It takes a request payload in JSON format, which should
contain; title, description and conclusion of an entry to be modify. The
entry Id should be append to the end of request url with a preceeding
forward slash (eg. /api/v1/entries</id>). Request to this endpoint
will use HTTP verb, "PUT".

If Id is not a valid non-negative integer value, the endpoint returns
error object and a HTTP status code; 400, to the requesting client.

Title and description to overwrite the existing ones in the entry are
required. If they are not supplied, the endpoint responds to the request
with HTTP status code; 400 and a JSON response data containing error
object.

If Id and request payload are valid, and an entry object found at the
position of the Id in entry list, the endpoint returns JSON object of
success message and a HTTP status code; 200, to the client. Else, the
endpoint returns a "Not found" error message and a HTTP status code;
404, to the client.